Refactor model construction
