+++ isoquant_lib/graph_based_model_construction.py
@@ -5,62 +5,34 @@
 # See file LICENSE for details.
 ############################################################################
 
-import copy
 import logging
 from collections import defaultdict
-from functools import cmp_to_key
-from enum import unique, Enum
-from typing import Callable, Dict, List, Optional, Tuple, TYPE_CHECKING
 
-from .common import (
-    TranscriptNaming,
-    cmp,
-    get_exons,
-    intersection_len,
-    interval_len,
-    junctions_from_blocks,
-)
+from .common import junctions_from_blocks
 from .assignment_io import ReadAssignmentType
-from .gene_info import GeneInfo, StrandDetector, TranscriptModel, TranscriptModelType
-from .intron_graph import IntronGraph, TerminalVertex
+from .gene_info import TranscriptModelType
 from .isoform_assignment import (
-    is_matching_assignment,
     match_subtype_to_str_with_additional_info,
-    MatchEventSubtype,
     ReadAssignment,
     MatchClassification,
-    IsoformMatch
+    IsoformMatch,
 )
-from .long_read_assigner import LongReadAssigner
-from .transcript_splice_site_corrector import (
-    count_deletions_for_splice_site_locations,
-    correct_splice_site_errors,
-    generate_updated_exon_list,
-    SUPPORTED_STRANDS,
-)
-from .long_read_profiles import CombinedProfileConstructor
 from .polya_finder import PolyAInfo
-from .terminal_peaks import detect_peaks, get_polya_model, get_tss_model
 from .transcript_to_gene_joiner import TranscriptToGeneJoiner
-
-if TYPE_CHECKING:
-    from xgboost import XGBClassifier
-
+from .model_construction.context import ModelStore, ModelConstructionContext
+# Re-exported for isoquant.py's `from ...graph_based_model_construction import StrandnessReportingLevel`.
+from .model_construction.context import StrandnessReportingLevel  # noqa: F401
+from .model_construction.fl_graph_constructor import FLGraphConstructor
+from .model_construction.assignment_based_constructor import AssignmentBasedConstructor
+from .model_construction.end_processor import TranscriptEndProcessor
+from .model_construction.model_filter import ModelFilter
+from .model_construction.read_assigner import ConstructionReadAssigner
+from .model_construction.model_read_counter import ModelReadCounter
 
 logger = logging.getLogger('IsoQuant')
 
 
-@unique
-class StrandnessReportingLevel(Enum):
-    only_canonical = 1
-    only_stranded = 2
-    all = 3
-    auto = 10
-
-
 class GraphBasedModelConstructor:
-    detected_known_isoforms = set()
-    extended_transcript_ids = set()
 
     def __init__(self, gene_info, chr_record, args, transcript_counter, gene_counter, id_distributor,
                  grouping_strategy_names=None,
@@ -68,305 +40,81 @@ class GraphBasedModelConstructor:
                  string_pools=None,
                  polya_predictions=None,
                  tss_predictions=None):
-        self.gene_info = gene_info
-        self.chr_record = chr_record
-        self.args = args
-        self.id_distributor = id_distributor
-        self.string_pools = string_pools
-        # Predicted polyA / TSS sites for this gene (genomic positions), reused
-        # from the per-gene terminal-position counters to refine intron-graph
-        # terminal vertices.
-        self.polya_predictions = polya_predictions
-        self.tss_predictions = tss_predictions
-        self.grouping_strategy_names = grouping_strategy_names if grouping_strategy_names else []
-        self.use_technical_replicas = use_technical_replicas
-        # Find file_name group index for technical replicas check
-        self.file_name_group_idx = self.grouping_strategy_names.index("file_name") if "file_name" in self.grouping_strategy_names else -1
-
-        # Constructed (novel) model ends are always refined to the trained
-        # polyA/TSS peaks (see correct_novel_transcript_ends). Runs with or
-        # without an annotation (needs only the shipped detector + the model's
-        # own reads); never creates or drops models.
-        # Alternative-end NIC *creation* (graph-level reclassification +
-        # post-pass known->NIC) requires an annotation to refine against, so it
-        # is gated on --genedb. Keeps the with-genedb behaviour unchanged.
-        self.create_nics = bool(getattr(args, 'genedb', None))
-        # TSS model is used (for polishing and for NIC creation) only with
-        # full-length evidence; read starts are unreliable otherwise.
-        self.use_tss_model = bool(getattr(args, 'fl_data', False))
-        # Opt-in (--novel_apa, default off): also create alternative-end isoforms
-        # for novel (non-reference) transcripts; works with or without --genedb.
-        self.novel_apa = bool(getattr(args, 'novel_apa', False))
-
-        self.strand_detector = StrandDetector(self.chr_record)
-        self.intron_genes = defaultdict(set)
-        self.set_gene_properties()
-
-        self.intron_graph = None
-        self.path_processor = None
-        self.path_storage = None
-
-        self.known_isoforms_in_graph = {}
-        self.known_introns = set()
-        self.known_isoforms_in_graph_ids = {}
-        self.assigner = LongReadAssigner(self.gene_info, self.args, self.string_pools)
-        self.profile_constructor = CombinedProfileConstructor(self.gene_info, self.args)
-
-        self.transcript_model_storage = []
-        self.transcript_read_ids = defaultdict(list)
-        self.transcript_counter = transcript_counter
-        self.gene_counter = gene_counter
-        self.internal_counter = defaultdict(int)
-        self.read_assignment_counts = defaultdict(int)
+        # Read-only inputs + intron graph + the shared ModelStore live on the ctx;
+        # the extracted construction stages each receive this ctx.
+        store = ModelStore(gene_info)
+        self.ctx = ModelConstructionContext(
+            gene_info, chr_record, args, id_distributor, string_pools,
+            polya_predictions, tss_predictions, store,
+            grouping_strategy_names=grouping_strategy_names,
+            use_technical_replicas=use_technical_replicas)
+        ctx = self.ctx
+        self.store = store
+        # Inputs used by this orchestrator itself (gene_info is also read by
+        # parallel_workers; the rest drive process()/compare_models_with_known).
+        self.gene_info = ctx.gene_info
+        self.args = ctx.args
+        self.string_pools = ctx.string_pools
+        self.id_distributor = ctx.id_distributor
+        self.assigner = ctx.assigner
+        self.profile_constructor = ctx.profile_constructor
         self.transcript2transcript = []
-        # read_id -> transcript id the read was bound to during construction
-        # (via save_assigned_read). Lets build_model_read_assignments tell
-        # construction-supported reads (kept unique/umd) apart from reads that
-        # only reach the re-profiling step (classified honestly).
-        self.construction_assignment: Dict[str, str] = {}
-        # known transcript id -> reads that were used to build it as a non-FL model
-        # (construct_nonfl_isoforms). Includes reads already bound to a novel FL model;
-        # resolve_read_ownership_conflicts uses this after filtering to hand any read
-        # still shared with a surviving novel back to that novel.
-        self.nonfl_known_reads: Dict[str, List[ReadAssignment]] = {}
-        # Honest per-read ReadAssignments against the final model set, produced by
-        # build_model_read_assignments; fed to the discovered-model counters and,
-        # when read2transcripts is on, to the model read_info printer.
-        self.model_read_assignments: List[ReadAssignment] = []
+
+        # Construction stages, each operating on the shared ctx / ctx.store.
+        # Stage 1: full-length isoforms from intron-graph paths.
+        self.fl_constructor = FLGraphConstructor(ctx)
+        # Stage 2: models from per-read reference assignments.
+        self.assignment_constructor = AssignmentBasedConstructor(ctx)
+        # Stage 3: polyA/TSS end refinement + alternative-end NIC discovery.
+        self.end_processor = TranscriptEndProcessor(ctx)
+        # Stage 4a: structural filtering + splice-site correction (uses stage 3).
+        self.model_filter = ModelFilter(ctx, self.end_processor)
+        # Stage 4b: construction-phase read (re)assignment + ownership resolution.
+        self.read_assigner = ConstructionReadAssigner(ctx)
+        # Stage 4c: final honest per-read assignment + counting + read_info.
+        self.model_read_counter = ModelReadCounter(ctx, transcript_counter, gene_counter)
+
+    @property
+    def transcript_model_storage(self):
+        # Public accessor for the final model list (read by parallel_workers /
+        # printers); the list itself lives on the shared ModelStore.
+        return self.store.transcript_model_storage
+
+    @property
+    def model_read_assignments(self):
+        # Honest per-read ReadAssignments against the final model set (read by
+        # parallel_workers for the model read_info output); produced by stage 4c.
+        return self.model_read_counter.model_read_assignments
 
     def get_transcript_id(self):
         return self.id_distributor.increment()
 
-    def set_gene_properties(self):
-        intron_strands_dicts = defaultdict(lambda: defaultdict(int))
-        self.intron_genes = defaultdict(set)
-        for t_id, introns in self.gene_info.all_isoforms_introns.items():
-            strand = self.gene_info.isoform_strands[t_id]
-            gene_id = self.gene_info.gene_id_map[t_id]
-            for intron in introns:
-                intron_strands_dicts[intron][strand] += 1
-                self.intron_genes[intron].add(gene_id)
-
-        for intron in intron_strands_dicts.keys():
-            if len(intron_strands_dicts[intron].keys()) == 1:
-                # intron has a single strand
-                self.strand_detector.set_strand(intron, list(intron_strands_dicts[intron].keys())[0])
-            else:
-                self.strand_detector.set_strand(intron)
-
-    def select_reference_gene(self, transcript_introns, transcript_range, transcript_strand):
-        if self.gene_info.empty():
-            return None
-        gene_counts = defaultdict(int)
-        for intron in transcript_introns:
-            if intron not in self.intron_genes:
-                continue
-            for g_id in self.intron_genes[intron]:
-                gene_counts[g_id] += 1
-
-        ordered_genes = sorted(gene_counts.items(), key=lambda x: (x[1], x[0]), reverse=True)
-        for g in ordered_genes:
-            gene_id = g[0]
-            if transcript_strand == '.' or self.gene_info.gene_strands[gene_id] == transcript_strand:
-                return gene_id
-
-        return None
-
     def process(self, read_assignment_storage):
-        self.intron_graph = IntronGraph(self.args, self.gene_info, read_assignment_storage,
-                                        polya_predictions=self.polya_predictions,
-                                        tss_predictions=self.tss_predictions)
-        self.path_processor = IntronPathProcessor(self.args, self.intron_graph)
-        self.path_storage = IntronPathStorage(self.args, self.path_processor)
-        self.path_storage.fill(read_assignment_storage)
-        self.known_isoforms_in_graph = self.get_known_spliced_isoforms(self.gene_info)
-        self.known_introns = set(self.gene_info.intron_profiles.features)
+        self.ctx.build_graph(read_assignment_storage)
 
-        for intron_path, isoform_id in self.known_isoforms_in_graph.items():
-            self.known_isoforms_in_graph_ids[isoform_id] = intron_path
+        self.fl_constructor.construct_fl_isoforms()
+        self.assignment_constructor.construct_assignment_based_isoforms(read_assignment_storage)
 
-        self.construct_fl_isoforms()
-        self.construct_assignment_based_isoforms(read_assignment_storage)
-
-        self.pre_filter_transcripts()
-        self.assign_reads_to_models(read_assignment_storage)
-        self.filter_transcripts()
-        self.correct_transcript_splice_sites()
+        self.model_filter.pre_filter_transcripts()
+        self.read_assigner.assign_reads_to_models(read_assignment_storage)
+        self.model_filter.filter_transcripts()
+        self.model_filter.correct_transcript_splice_sites()
         # reassign reads
-        self.assign_reads_to_models(read_assignment_storage)
+        self.read_assigner.assign_reads_to_models(read_assignment_storage)
         # hand reads shared between an FL-path model and an assignment-based known back
         # to the FL-path model (novel-vs-known and known-vs-known), dropping empty knowns
-        self.resolve_read_ownership_conflicts()
+        self.read_assigner.resolve_read_ownership_conflicts()
 
-        transcript_joiner = TranscriptToGeneJoiner(self.transcript_model_storage, self.gene_info)
-        self.transcript_model_storage = transcript_joiner.join_transcripts()
-        self.build_model_read_assignments(read_assignment_storage)
+        transcript_joiner = TranscriptToGeneJoiner(self.store.transcript_model_storage, self.gene_info)
+        self.store.transcript_model_storage = transcript_joiner.join_transcripts()
+        self.model_read_counter.build_model_read_assignments(read_assignment_storage)
 
         if self.args.sqanti_output:
             self.compare_models_with_known()
 
-    def _log_uniqueknown_on_novel(self, transcript2type: Dict[str, TranscriptModelType]) -> None:
-        # Diagnostic: how often a read the reference step assigned uniquely to a
-        # known isoform ends up on a *novel* model. After the re-assignment gate
-        # this can only happen while the read is consumed building that model
-        # (FL path / alt-end), so the count isolates the construction-phase drag.
-        dragged = 0
-        for t_id, reads in self.transcript_read_ids.items():
-            if transcript2type.get(t_id) == TranscriptModelType.known:
-                continue
-            for a in reads:
-                x = self._reference_unique_known_isoform(a)
-                if x is not None and x != t_id:
-                    dragged += 1
-                    logger.debug("Read %s uniquely assigned to known %s now on novel model %s" %
-                                 (a.read_id, x, t_id))
-        if dragged:
-            logger.debug("%d uniquely-known reads bound to novel models (construction-phase)" % dragged)
-
-    def _single_model_assigner(self, model, cache):
-        # Cache one (assigner, profile_constructor) per model over a single-isoform
-        # GeneInfo, using the graph vertex-collapse tolerance (graph_clustering_distance)
-        # as the junction delta so a construction read whose splice sites the intron
-        # graph snapped within that reach still matches its model, instead of being
-        # rejected at the tighter assignment delta. quick_mode=False for real events.
-        entry = cache.get(model.transcript_id)
-        if entry is None:
-            gi = GeneInfo.from_models([model], self.args.delta)
-            entry = (LongReadAssigner(gi, self._graph_tolerance_args, self.string_pools, quick_mode=False),
-                     CombinedProfileConstructor(gi, self._graph_tolerance_args))
-            cache[model.transcript_id] = entry
-        return entry
-
-    def _model_read_assignment_single(self, source, transcript_id, model_by_id, cache):
-        # A read that supports this model must count toward it. Reuse the source
-        # match when the read was reference unique/umd on a *known* model (real
-        # classification + events, no re-profiling). Otherwise re-profile against
-        # the single model at the graph tolerance; keep the result when it is a
-        # clean consistent match (real classification + events), else force the
-        # read onto the model as unique_minor_difference (events dropped) so it
-        # still counts as unique and points at the model it built.
-        model = model_by_id[transcript_id]
-        if (model.transcript_type == TranscriptModelType.known and
-                source.assignment_type in (ReadAssignmentType.unique,
-                                           ReadAssignmentType.unique_minor_difference)):
-            reuse_match = next((m for m in source.isoform_matches
-                                if m.assigned_transcript == transcript_id), None)
-            if reuse_match is not None:
-                return ReadAssignment(source.read_id, source.assignment_type,
-                                      self.string_pools, match=[reuse_match])
-        assigner, profile_constructor = self._single_model_assigner(model, cache)
-        profile = profile_constructor.construct_profiles(source.corrected_exons, source.polya_info, [])
-        ra = assigner.assign_to_isoform(source.read_id, profile)
-        if (ra.assignment_type.is_consistent() and ra.isoform_matches and
-                ra.isoform_matches[0].assigned_transcript == transcript_id):
-            return ra
-        forced_match = IsoformMatch(MatchClassification.full_splice_match, self.string_pools,
-                                    assigned_gene=model.gene_id, assigned_transcript=transcript_id)
-        return ReadAssignment(source.read_id, ReadAssignmentType.unique_minor_difference,
-                              self.string_pools, match=[forced_match])
-
-    def _build_discarded_assignment(self, read_id, gene, surviving_genes):
-        # A read whose reference-unique known isoform was dropped: no transcript.
-        # It still belongs to its gene, so it counts toward that gene when the gene
-        # has a surviving model (gene_assignment_type=unique makes the gene counter
-        # count it); otherwise it is not assigned anywhere.
-        if gene is not None and gene in surviving_genes:
-            match = IsoformMatch(MatchClassification.genic, self.string_pools, assigned_gene=gene)
-            ra = ReadAssignment(read_id, ReadAssignmentType.discarded, self.string_pools, match=[match])
-            ra.gene_assignment_type = ReadAssignmentType.unique
-            return ra
-        return ReadAssignment(read_id, ReadAssignmentType.discarded, self.string_pools)
-
-    def build_model_read_assignments(self, read_assignments):
-        """Produce one honest ReadAssignment per read against the final model set,
-        feed each to the discovered transcript & gene counters (which apply the
-        per-type counting strategy) and collect them for the model read_info
-        printer. Replaces the artificial forward_counts, so read2transcripts comes
-        for free. Categories (construction checked first, mirroring the pipeline
-        where construction-bound reads skip the re-profiling gate):
-
-        1. construction member  -> re-profile vs the single model it built at the
-           graph tolerance, forced onto it as unique/umd (it must count toward the
-           model it contributed to).
-        2. reference-unique member on a *kept* known isoform -> reuse source match.
-        3. discarded (reference-unique to a *dropped* known isoform) -> no transcript;
-           counts toward the gene only if that gene still has a surviving model.
-        4. leftover -> assign vs the full model set (quick_mode=False): honest
-           unique/umd/ambiguous/inconsistent, kept regardless of consistency.
-        """
-        self.model_read_assignments = []
-        # Junction delta for the single-model (construction/known) re-profiling:
-        # the intron graph's vertex-collapse tolerance, so reads snapped within it
-        # match their model. Never below the assignment delta.
-        self._graph_tolerance_args = copy.copy(self.args)
-        _gcd = getattr(self.args, "graph_clustering_distance", None)
-        self._graph_tolerance_args.delta = max(self.args.delta, _gcd) if _gcd else self.args.delta
-        transcript2gene = {t.transcript_id: t.gene_id for t in self.transcript_model_storage}
-        transcript2type = {t.transcript_id: t.transcript_type for t in self.transcript_model_storage}
-        self._log_uniqueknown_on_novel(transcript2type)
-
-        model_gene_info = GeneInfo.from_models(self.transcript_model_storage, self.args.delta)
-        model_by_id = {t.transcript_id: t for t in self.transcript_model_storage}
-        model_ids = set(model_by_id.keys())
-        surviving_genes = set(transcript2gene.values())
-
-        single_cache = {}
-        full_assigner = None
-        full_profile_constructor = None
-
-        for source in read_assignments:
-            read_id = source.read_id
-            constr_tid = self.construction_assignment.get(read_id)
-            if constr_tid is not None and constr_tid in model_ids:
-                ra = self._model_read_assignment_single(source, constr_tid, model_by_id, single_cache)
-            else:
-                known_iso = self._reference_unique_known_isoform(source)
-                if known_iso is not None and known_iso in model_ids:
-                    ra = self._model_read_assignment_single(source, known_iso, model_by_id, single_cache)
-                elif known_iso is not None:
-                    gene = self.gene_info.gene_id_map.get(known_iso)
-                    ra = self._build_discarded_assignment(read_id, gene, surviving_genes)
-                elif model_ids:
-                    if full_assigner is None:
-                        full_assigner = LongReadAssigner(model_gene_info, self.args,
-                                                         self.string_pools, quick_mode=False)
-                        full_profile_constructor = CombinedProfileConstructor(model_gene_info, self.args)
-                    profile = full_profile_constructor.construct_profiles(source.corrected_exons,
-                                                                          source.polya_info, [])
-                    ra = full_assigner.assign_to_isoform(read_id, profile)
-                else:
-                    ra = ReadAssignment(read_id, ReadAssignmentType.noninformative, self.string_pools)
-
-            ra.gene_info = model_gene_info
-            self._copy_read_aux(ra, source)
-            self.model_read_assignments.append(ra)
-            self.transcript_counter.add_read_info(ra)
-            self.gene_counter.add_read_info(ra)
-
-        self.transcript_counter.add_confirmed_features([m.transcript_id for m in self.transcript_model_storage])
-        self.gene_counter.add_confirmed_features({transcript2gene[m.transcript_id] for m in self.transcript_model_storage})
-
-    @staticmethod
-    def _copy_read_aux(ra: ReadAssignment, source: ReadAssignment) -> None:
-        # Carry every read_info aux field over from the source reference
-        # assignment so the model read_info output matches the main read_info
-        # format. ra shares the source's string pools, so the interned integer
-        # ids (chr/barcode/umi/read_group) can be copied directly.
-        ra.chr_id_int = source.chr_id_int
-        ra.strand = source.strand
-        ra.exons = source.exons
-        ra.corrected_exons = source.corrected_exons
-        ra.polyA_found = source.polyA_found
-        ra.cage_found = source.cage_found
-        ra.polya_info = source.polya_info
-        ra.barcode_id = source.barcode_id
-        ra.umi_id = source.umi_id
-        ra.read_group_ids = source.read_group_ids
-        ra.additional_attributes = source.additional_attributes
-
     def compare_models_with_known(self):
         if not self.gene_info.all_isoforms_exons:
-            for model in self.transcript_model_storage:
+            for model in self.store.transcript_model_storage:
                 # create intergenic
                 assignment = ReadAssignment(model.transcript_id,
                                             ReadAssignmentType.intergenic,
@@ -393,11 +141,11 @@ class GraphBasedModelConstructor:
             return
 
         gene_to_model_dict = defaultdict(list)
-        for model in self.transcript_model_storage:
+        for model in self.store.transcript_model_storage:
             gene_to_model_dict[model.gene_id].append(model.transcript_id)
 
         self.transcript2transcript = []
-        for model in self.transcript_model_storage:
+        for model in self.store.transcript_model_storage:
             if model.transcript_type == TranscriptModelType.known:
                 continue
             if model.strand == "-":
@@ -447,1253 +195,3 @@ class GraphBasedModelConstructor:
             model.add_additional_attribute("similar_reference_id", assigned_transcript_id)
             model.add_additional_attribute("alternatives", event_string)
             self.transcript2transcript.append(assignment)
-
-    def pre_filter_transcripts(self):
-        internal_count_values = []
-        for model in self.transcript_model_storage:
-            if len(model.exon_blocks) <= 2:
-                internal_count_values.append(self.internal_counter[model.transcript_id])
-
-        internal_count_values = sorted(internal_count_values, reverse=True)
-        coverage_cutoff = self.args.min_novel_count
-        # dirty hack to avoid slow assignment in chrM
-        if len(internal_count_values) > 50:
-            coverage_cutoff = internal_count_values[50]
-
-        filtered_storage = []
-        for model in self.transcript_model_storage:
-            if len(model.exon_blocks) > 2:
-                filtered_storage.append(model)
-                continue
-
-            if (model.transcript_type != TranscriptModelType.known and
-                    self.internal_counter[model.transcript_id] < coverage_cutoff):
-                self.delete_from_storage(model.transcript_id)
-                continue
-
-            if (model.transcript_type != TranscriptModelType.known and
-                    self.mapping_quality(model.transcript_id) < self.args.simple_models_mapq_cutoff):
-                self.delete_from_storage(model.transcript_id)
-                continue
-            filtered_storage.append(model)
-
-        self.transcript_model_storage = filtered_storage
-
-    def delete_from_storage(self, transcript_id):
-        for a in self.transcript_read_ids[transcript_id]:
-            self.read_assignment_counts[a.read_id] -= 1
-        del self.transcript_read_ids[transcript_id]
-        del self.internal_counter[transcript_id]
-
-    def resolve_read_ownership_conflicts(self) -> None:
-        # construct_nonfl_isoforms builds an assignment-based known model and re-binds
-        # (save_assigned_read overwrites construction_assignment) reads that an FL path
-        # already bound to another model. Hand such reads back to the FL-path model when
-        # the assignment-based known would otherwise steal the FL model's only support:
-        #   - a novel FL model (novel-vs-known): the novel is a distinct discovered
-        #     transcript and always wins its reads;
-        #   - an FL-path known left with zero owned reads (known-vs-known): a near-identical
-        #     sibling of the assignment-based known (their intron chains differ by a few bp
-        #     at one splice site, collapsed in the intron graph) stole all its reads, so it
-        #     is a zero-count phantom -> reclaim them. An FL known that still owns reads is
-        #     a legitimate co-expressed isoform and is left alone (no sibling consolidation).
-        # After filtering, drop assignment-based knowns left with no reads.
-        if not self.nonfl_known_reads:
-            return
-        # reads currently owned by each model (its construction_assignment target)
-        owned = defaultdict(int)
-        for mid in self.construction_assignment.values():
-            owned[mid] += 1
-        # read_id -> a surviving winner model that should reclaim it: a novel FL model,
-        # or an FL-path known that is currently a zero-count phantom (0 owned reads).
-        read_to_fl: Dict[str, str] = {}
-        for model in self.transcript_model_storage:
-            if model.transcript_id in self.nonfl_known_reads:
-                continue  # assignment-based known: a potential loser, never a winner
-            if (model.transcript_type == TranscriptModelType.known and
-                    owned.get(model.transcript_id, 0) > 0):
-                continue  # FL-path known that still owns reads: co-expressed, leave it
-            for a in self.transcript_read_ids[model.transcript_id]:
-                read_to_fl.setdefault(a.read_id, model.transcript_id)
-
-        surviving_ids = {m.transcript_id for m in self.transcript_model_storage}
-        emptied_knowns = []
-        for known_id, reads in self.nonfl_known_reads.items():
-            if known_id not in surviving_ids:
-                continue
-            for a in reads:
-                if self.construction_assignment.get(a.read_id) != known_id:
-                    # read no longer owned by this known (e.g. moved to an alt-end NIC)
-                    continue
-                fl_id = read_to_fl.get(a.read_id)
-                if fl_id is None or fl_id == known_id:
-                    continue
-                # still shared with a surviving FL-path model -> the FL model wins
-                self.construction_assignment[a.read_id] = fl_id
-                if a in self.transcript_read_ids[known_id]:
-                    self.transcript_read_ids[known_id].remove(a)
-                    self.internal_counter[known_id] -= 1
-                    self.read_assignment_counts[a.read_id] -= 1
-            if self.internal_counter[known_id] <= 0:
-                emptied_knowns.append(known_id)
-
-        if emptied_knowns:
-            drop = set(emptied_knowns)
-            for known_id in drop:
-                self.delete_from_storage(known_id)
-            self.transcript_model_storage = [m for m in self.transcript_model_storage
-                                             if m.transcript_id not in drop]
-            logger.debug("resolve_read_ownership_conflicts: dropped %d empty assignment-based knowns"
-                         % len(drop))
-
-    def filter_transcripts(self):
-        pre_filtered_storage = []
-        to_substitute = self.detect_similar_isoforms(self.transcript_model_storage)
-
-        for model in self.transcript_model_storage:
-            if model.transcript_type == TranscriptModelType.known:
-                pre_filtered_storage.append(model)
-                continue
-            # check coverage
-            component_coverage = self.intron_graph.get_max_component_coverage(model.intron_path)
-            if component_coverage == 0 or len(model.intron_path) == 0 or \
-                    (len(model.intron_path) == 1 and self.intron_graph.is_monointron(model.intron_path[0])):
-                component_coverage = self.intron_graph.get_overlapping_component_max_coverage((model.get_start(), model.get_end()))
-                novel_isoform_cutoff = max(self.args.min_novel_count,
-                                           self.args.min_mono_count_rel * component_coverage)
-            else:
-                novel_isoform_cutoff = max(self.args.min_novel_count,
-                                           self.args.min_novel_count_rel * component_coverage)
-
-            if model.transcript_id in to_substitute:
-                #logger.debug("Novel model %s has a similar isoform %s" % (model.transcript_id, to_substitute[model.transcript_id]))
-                self.delete_from_storage(model.transcript_id)
-                continue
-
-            if self.internal_counter[model.transcript_id] < novel_isoform_cutoff:
-                #logger.debug("Novel model %s has coverage %d < %.2f, component cov = %d" % (model.transcript_id,
-                #                                                        self.internal_counter[model.transcript_id],
-                #                                                        novel_isoform_cutoff, component_coverage))
-                self.delete_from_storage(model.transcript_id)
-                continue
-
-            if len(model.exon_blocks) <= 2:
-                mapq = self.mapping_quality(model.transcript_id)
-                #logger.debug("Novel model %s has quality %.2f" % (model.transcript_id, mapq))
-                if mapq < self.args.simple_models_mapq_cutoff:
-                    #logger.debug("Novel model %s has poor quality" % model.transcript_id)
-                    self.delete_from_storage(model.transcript_id)
-                    continue
-
-            # TODO: correct ends for known
-            self.correct_novel_transcript_ends(model, self.transcript_read_ids[model.transcript_id])
-            self._mark_terminal_confirmation(model)
-            pre_filtered_storage.append(model)
-
-
-        filtered_storage = []
-        to_substitute = self.detect_similar_isoforms(pre_filtered_storage, remove_unconfirmed=True)
-
-        for model in pre_filtered_storage:
-            if model.transcript_type == TranscriptModelType.known:
-                filtered_storage.append(model)
-                continue
-
-            if model.transcript_id in to_substitute:
-                #logger.debug("Novel model %s has a similar isoform %s" % (model.transcript_id, to_substitute[model.transcript_id]))
-                self.delete_from_storage(model.transcript_id)
-                continue
-
-            filtered_storage.append(model)
-
-        self.transcript_model_storage = filtered_storage
-        if self.create_nics or self.novel_apa:
-            self._add_known_alternative_end_models()
-            self._drop_duplicate_alt_end_models()
-
-    def correct_transcript_splice_sites(self) -> None:
-        """Shift novel transcript-model splice sites onto canonical motifs when a
-        consistent cluster of read deletions next to a junction indicates the
-        aligner mis-placed it. Known reference junctions come from the annotation
-        and are left untouched."""
-        if getattr(self.args, "no_splice_site_correction", False):
-            return
-        for model in self.transcript_model_storage:
-            if model.transcript_type == TranscriptModelType.known:
-                continue
-            corrected_exons = self._correct_one_model_splice_sites(
-                model, self.transcript_read_ids[model.transcript_id])
-            if corrected_exons and corrected_exons != model.exon_blocks and \
-                    self._valid_exon_chain(corrected_exons):
-                logger.debug("Corrected splice sites for %s: %s -> %s" %
-                             (model.transcript_id, model.exon_blocks, corrected_exons))
-                model.exon_blocks = corrected_exons
-
-    def _correct_one_model_splice_sites(self, model, reads):
-        """Return a corrected exon-block list for a single model, or None."""
-        exons = model.exon_blocks
-        if len(exons) < 2 or not reads or model.strand not in SUPPORTED_STRANDS:
-            return None
-
-        splice_site_cases = {}
-        for read in reads:
-            if not read.exons:
-                continue
-            count_deletions_for_splice_site_locations(
-                read.exons[0][0], read.exons[-1][1], read.boundary_deletions or [],
-                exons, splice_site_cases)
-
-        if not splice_site_cases:
-            return None
-        locations_with_errors = correct_splice_site_errors(
-            splice_site_cases, self.chr_record, model.strand)
-        if not locations_with_errors:
-            return None
-        return generate_updated_exon_list(splice_site_cases, locations_with_errors, exons)
-
-    @staticmethod
-    def _valid_exon_chain(exons) -> bool:
-        """Guard against a correction that inverts an exon or collides with a
-        neighbour (introns must keep at least one base)."""
-        for i, (start, end) in enumerate(exons):
-            if start > end:
-                return False
-            if i > 0 and start <= exons[i - 1][1]:
-                return False
-        return True
-
-    def mapping_quality(self, transcript_id):
-        mapq = 0
-        for a in self.transcript_read_ids[transcript_id]:
-            mapq += a.mapping_quality
-        return mapq / len(self.transcript_read_ids[transcript_id])
-
-    def detect_similar_isoforms(self, model_storage, remove_unconfirmed=False):
-        to_substitute = {}
-        for model in model_storage:
-            if len(model.exon_blocks) <= 2 or model.transcript_id in to_substitute:
-                continue
-            transcript_model_gene_info = GeneInfo.from_models([model], self.args.delta)
-            assigner = LongReadAssigner(transcript_model_gene_info, self.args, self.string_pools)
-            profile_constructor = CombinedProfileConstructor(transcript_model_gene_info, self.args)
-
-            for m in model_storage:
-                if m.transcript_type == TranscriptModelType.known or m.transcript_id == model.transcript_id or \
-                        m.transcript_id in to_substitute or len(m.exon_blocks) == 1 or not m.intron_path or \
-                        len(m.exon_blocks) > len(model.exon_blocks):
-                    continue
-
-                if m.intron_path[0][0] == TerminalVertex.polyt:
-                    polya_info = PolyAInfo(-1, m.intron_path[0][1], -1, -1)
-                elif m.intron_path[-1][0] == TerminalVertex.polya:
-                    polya_info = PolyAInfo(m.intron_path[-1][1], -1, -1, -1)
-                else:
-                    polya_info = PolyAInfo(-1, -1, -1, -1)
-                combined_profile = profile_constructor.construct_profiles(m.exon_blocks, polya_info, [])
-                assignment = assigner.assign_to_isoform(m.transcript_id, combined_profile)
-
-                if is_matching_assignment(assignment):
-                    to_substitute[m.transcript_id] = model.transcript_id
-                elif (remove_unconfirmed and
-                      assignment.assignment_type in (ReadAssignmentType.unique_minor_difference,
-                                                     ReadAssignmentType.inconsistent_non_intronic) and
-                      not (m.polya_confirmed or (self.use_tss_model and m.tss_confirmed))):
-                    # m is a terminal/ISM sub-structure of the longer model: its intron chain is a
-                    # subset and only the terminal exons differ (unique_minor_difference with an
-                    # ISM/elongation event, or inconsistent_non_intronic), and its polyA/TSS
-                    # terminus is not confirmed by a read-terminus peak -> truncation artifact, not
-                    # a real alternative-end transcript. A genuinely different intron chain gives
-                    # plain `inconsistent` and is left alone. Remove it (reads reassigned to model).
-                    to_substitute[m.transcript_id] = model.transcript_id
-
-        return to_substitute
-
-    def get_known_spliced_isoforms(self, gene_info, s="known"):
-        known_isoforms = {}
-        for isoform_id in gene_info.all_isoforms_introns:
-            isoform_introns = gene_info.all_isoforms_introns[isoform_id]
-            intron_path = self.path_processor.thread_introns(isoform_introns)
-            if not intron_path:
-                #logger.debug("== No path found for %s isoform %s: %s" % (s, isoform_id, gene_info.all_isoforms_introns[isoform_id]))
-                continue
-            #logger.debug("== Path found for %s isoform %s: %s" % (s, isoform_id, gene_info.all_isoforms_introns[isoform_id]))
-            known_isoforms[tuple(intron_path)] = isoform_id
-        return known_isoforms
-
-    def save_assigned_read(self, read_assignment, transcript_id):
-        read_id = read_assignment.read_id
-        self.transcript_read_ids[transcript_id].append(read_assignment)
-        self.internal_counter[transcript_id] += 1
-        self.read_assignment_counts[read_id] += 1
-        self.construction_assignment[read_id] = transcript_id
-
-    def _move_read_to_model(self, read_assignment: ReadAssignment,
-                            source_id: str, target_id: str) -> None:
-        # Move one read's construction membership from source_id to target_id (an
-        # alternative-end NIC derived from it). ReadAssignment has no __eq__, so
-        # list.remove strips the exact object; net read_assignment_counts is
-        # unchanged (-1 here, +1 in save_assigned_read). Keeps the counting
-        # invariants so the NIC is no longer a zero-read phantom.
-        self.transcript_read_ids[source_id].remove(read_assignment)
-        self.internal_counter[source_id] -= 1
-        self.read_assignment_counts[read_assignment.read_id] -= 1
-        self.save_assigned_read(read_assignment, target_id)
-
-    def _reference_isoform_for_path(self, assignment: ReadAssignment, path: tuple, intron_path: tuple,
-                                    transcript_range: Tuple[int, int]) -> Optional[str]:
-        # Reference isoform whose intron chain matches this path. It is reported
-        # as the annotated known UNLESS a *detected polyA* terminal vertex
-        # (TerminalVertex.polya / TerminalVertex.polyt) disagrees with the annotated end by more
-        # than apa_delta -> then the caller emits a novel-in-catalog isoform with
-        # the refined ends. A bare read_end / read_start (no polyA evidence, e.g.
-        # a degraded ONT terminus) never triggers reclassification, so known
-        # transcripts are not lost to noise.
-        if is_matching_assignment(assignment):
-            matched_reference_id = assignment.isoform_matches[0].assigned_transcript
-        elif intron_path in self.known_isoforms_in_graph:
-            matched_reference_id = self.known_isoforms_in_graph[intron_path]
-        else:
-            return None
-
-        if matched_reference_id is None or \
-                matched_reference_id not in self.gene_info.all_isoforms_exons:
-            return None
-
-        ref_exons = self.gene_info.all_isoforms_exons[matched_reference_id]
-        left_diff = abs(transcript_range[0] - ref_exons[0][0]) > self.args.apa_delta
-        right_diff = abs(transcript_range[1] - ref_exons[-1][1]) > self.args.apa_delta
-
-        # PolyA side: a detected polyA vertex (TerminalVertex.polyt on the genomic-left
-        # 3' end of a '-' transcript, TerminalVertex.polya on the genomic-right 3' end of
-        # a '+' transcript) that disagrees with the annotation -> alternative
-        # polyA NIC. Bare read termini never trigger here (degraded-end safety).
-        if path[0][0] == TerminalVertex.polyt and left_diff:
-            return None
-        if path[-1][0] == TerminalVertex.polya and right_diff:
-            return None
-
-        # TSS side (only with --fl_data): a *confirmed* TSS vertex (snapped to a TSS
-        # prediction, tss_left on the genomic-left 5' end of a '+' transcript /
-        # tss_right on the genomic-right 5' end of a '-' transcript) that disagrees
-        # with the annotation -> alternative-TSS NIC. Symmetric to the polyA block:
-        # a bare read_start / read_end never triggers, so ordinary 5' read-start
-        # scatter/extension does not spawn a NIC that steals the known's reads.
-        if self.use_tss_model:
-            strand = self.gene_info.isoform_strands.get(matched_reference_id, '.')
-            if strand == '+' and path[0][0] == TerminalVertex.tss_left and left_diff:
-                return None
-            if strand == '-' and path[-1][0] == TerminalVertex.tss_right and right_diff:
-                return None
-
-        return matched_reference_id
-
-    def construct_fl_isoforms(self):
-        # a minor trick to compare tuples of pairs, whose starting and terminating elements have different type
-        logger.debug("Total FL paths %d" % len(self.path_storage.fl_paths))
-        for path in sorted(self.path_storage.fl_paths,
-                           key=cmp_to_key(lambda x,y: cmp(x,y) if len(x)==len(y) else cmp(len(y), len(x)))):
-            # do not include terminal vertices
-            # logger.debug(">>> Considering path " + str(path))
-            intron_path = path[1:-1]
-            if not intron_path: continue
-            transcript_range = (path[0][1], path[-1][1])
-            novel_exons = get_exons(transcript_range, list(intron_path))
-            count = self.path_storage.paths[path]
-            new_transcript_id = TranscriptNaming.transcript_prefix + str(self.get_transcript_id())
-            # logger.debug("uuu %s: %s" % (new_transcript_id, str(novel_exons)))
-
-            reference_isoform = None
-            # check if new transcript matches a reference one
-            if intron_path[0][0] == TerminalVertex.polyt:
-                polya_info = PolyAInfo(-1, intron_path[0][1], -1, -1)
-            elif intron_path[-1][0] == TerminalVertex.polya:
-                polya_info = PolyAInfo(intron_path[-1][1], -1, -1, -1)
-            else:
-                polya_info = PolyAInfo(-1, -1, -1, -1)
-            combined_profile = self.profile_constructor.construct_profiles(novel_exons, polya_info, [])
-            assignment = self.assigner.assign_to_isoform(new_transcript_id, combined_profile)
-            # check that no serious contradiction occurs
-            # logger.debug("uuu Checking novel transcript %s: %s; assignment type %s" %
-            #             (new_transcript_id, str(novel_exons), str(assignment.assignment_type)))
-
-            if self.create_nics:
-                # use the path's goal-1-refined terminal vertices: keep the
-                # reference only when both ends agree with the annotation within
-                # apa_delta, otherwise fall through to the novel branch below to
-                # emit an alternative-end NIC built from the refined exons
-                reference_isoform = self._reference_isoform_for_path(assignment, path, intron_path, transcript_range)
-            elif is_matching_assignment(assignment):
-                reference_isoform = assignment.isoform_matches[0].assigned_transcript
-                # logger.debug("uuu Substituting with known isoform %s" % reference_isoform)
-            elif intron_path in self.known_isoforms_in_graph:
-                # path was not assigned to any known isoform but intron chain still matches
-                continue
-
-            new_model = None
-            if reference_isoform:
-                # adding FL reference isoform
-                if reference_isoform in GraphBasedModelConstructor.detected_known_isoforms:
-                    pass
-                elif count < self.args.min_known_count:
-                    pass # logger.debug("uuu Isoform %s has low coverage %d" % (reference_isoform, count))
-                else:
-                    new_model = self.transcript_from_reference(reference_isoform)
-                    GraphBasedModelConstructor.detected_known_isoforms.add(reference_isoform)
-                    #logger.debug("Adding known spliced isoform %s" % reference_isoform)
-                    #logger.debug("Annotated positions: %d, %d, %s" % (new_model.exon_blocks[0][0], new_model.exon_blocks[-1][1], new_model.strand))
-                    #logger.debug("Graph positions: %s, %s" % (str(path[0]), str(path[-1])))
-            else:
-                # adding FL novel isoform
-                # component_coverage = self.intron_graph.get_max_component_coverage(intron_path)
-                novel_isoform_cutoff = self.args.min_novel_count
-
-                has_polyt = path[0][0] == TerminalVertex.polyt
-                has_polya = path[-1][0] == TerminalVertex.polya
-                polya_site = has_polya or has_polyt
-                transcript_strand = self.strand_detector.get_strand(intron_path, has_polya, has_polyt)
-                transcript_clean_strand = self.strand_detector.get_clean_strand(intron_path)
-
-                #logger.debug("uuu Novel isoform %s has coverage: %d cutoff = %d, component cov = %d, max_coverage = %d"
-                #             % (new_transcript_id, count, novel_isoform_cutoff, component_coverage, self.intron_graph.max_coverage))
-                if count < novel_isoform_cutoff:
-                    # logger.debug("uuu Novel isoform %s has low coverage: %d\t%d" % (new_transcript_id, count, novel_isoform_cutoff))
-                    pass
-                elif (len(novel_exons) == 2 and
-                      ((self.args.require_monointronic_polya and not polya_site) or transcript_clean_strand == '.')):
-                    # logger.debug("uuu Avoiding single intron %s isoform: %d\t%s" % (new_transcript_id, count, str(path)))
-                    pass
-                elif ((self.args.report_canonical_strategy == StrandnessReportingLevel.only_canonical
-                       and transcript_clean_strand == '.')
-                      or (self.args.report_canonical_strategy == StrandnessReportingLevel.only_stranded
-                          and transcript_strand == '.')):
-                    logger.debug("Avoiding unreliable transcript with %d exons (strand cannot be detected)" % len(novel_exons))
-                    pass
-                else:
-                    if self.use_technical_replicas and self.file_name_group_idx >= 0:
-                        # Check if reads come from same file (technical replicates)
-                        read_assignments = self.path_storage.paths_to_reads[path]
-                        if read_assignments:
-                            file_groups = set([a.read_group[self.file_name_group_idx]
-                                             for a in read_assignments
-                                             if self.file_name_group_idx < len(a.read_group)])
-                            if len(file_groups) <= 1:
-                                #logger.debug("%s was suspended due to technical replicas check" % new_transcript_id)
-                                continue
-
-                    transcript_gene = self.select_reference_gene(intron_path, transcript_range, transcript_strand)
-                    if transcript_gene is None:
-                        transcript_gene = (TranscriptNaming.novel_gene_prefix + self.gene_info.chr_id +
-                                           "_" + str(self.get_transcript_id()))
-                    elif transcript_strand == '.':
-                        transcript_strand = self.gene_info.gene_strands[transcript_gene]
-
-                    if all(intron in self.known_introns for intron in intron_path):
-                        transcript_type = TranscriptModelType.novel_in_catalog
-                        id_suffix = TranscriptNaming.nic_transcript_suffix
-                    else:
-                        transcript_type = TranscriptModelType.novel_not_in_catalog
-                        id_suffix = TranscriptNaming.nnic_transcript_suffix
-
-                    new_model = TranscriptModel(self.gene_info.chr_id, transcript_strand,
-                                                new_transcript_id + ".%s" % self.gene_info.chr_id + id_suffix,
-                                                transcript_gene, novel_exons, transcript_type)
-                    new_model.intron_path = intron_path
-                    logger.debug("uuu Adding novel spliced isoform %s : %d\t%d" % (new_transcript_id, count, novel_isoform_cutoff))
-
-            if new_model:
-                self.transcript_model_storage.append(new_model)
-                for read_assignment in self.path_storage.paths_to_reads[path]:
-                    self.save_assigned_read(read_assignment, new_model.transcript_id)
-
-    def construct_assignment_based_isoforms(self, read_assignment_storage):
-        spliced_isoform_reads = defaultdict(list)
-        isoform_left_support = defaultdict(int)
-        isoform_right_support = defaultdict(int)
-        polya_sites = defaultdict(int)
-        mono_exon_isoform_reads = defaultdict(list)
-        mono_exon_isoform_coverage = {}
-        novel_mono_exon_reads = []
-
-        for read_assignment in read_assignment_storage:
-            if len(read_assignment.corrected_exons) <= 2 and \
-                    (read_assignment.multimapper or read_assignment.mapping_quality < self.args.simple_alignments_mapq_cutoff):
-                continue
-
-            if not read_assignment:
-                continue
-            if (len(read_assignment.corrected_exons) == 1 and
-                    read_assignment.polyA_found and not read_assignment.multimapper and
-                    (read_assignment.assignment_type.is_unassigned() or
-                     read_assignment.assignment_type.is_inconsistent())):
-                novel_mono_exon_reads.append(read_assignment)
-
-            if not read_assignment.assignment_type.is_unique():
-                continue
-
-            refrenence_isoform_id = read_assignment.isoform_matches[0].assigned_transcript
-            if refrenence_isoform_id in GraphBasedModelConstructor.detected_known_isoforms:
-                continue
-
-            events = read_assignment.isoform_matches[0].match_subclassifications
-            if any(e.event_type == MatchEventSubtype.mono_exon_match for e in events):
-                mono_exon_isoform_reads[refrenence_isoform_id].append(read_assignment)
-                assert len(self.gene_info.all_isoforms_introns[refrenence_isoform_id]) == 0
-                transcript_start = self.gene_info.all_isoforms_exons[refrenence_isoform_id][0][0]
-                transcript_end = self.gene_info.all_isoforms_exons[refrenence_isoform_id][-1][1]
-                t_len = transcript_end - transcript_start + 1
-
-                if refrenence_isoform_id not in mono_exon_isoform_coverage:
-                    mono_exon_isoform_coverage[refrenence_isoform_id] = [0 for _ in range(t_len)]
-                start = max(0, read_assignment.corrected_exons[0][0] - transcript_start)
-                end = min(t_len, read_assignment.corrected_exons[-1][1] - transcript_start + 1)
-                for i in range(start, end):
-                    mono_exon_isoform_coverage[refrenence_isoform_id][i] = 1
-
-                if self.gene_info.isoform_strands[refrenence_isoform_id] == '+':
-                    if any(x.event_type == MatchEventSubtype.correct_polya_site_right for x in events):
-                        polya_sites[refrenence_isoform_id] += 1
-                else:
-                    if any(x.event_type == MatchEventSubtype.correct_polya_site_left for x in events):
-                        polya_sites[refrenence_isoform_id] += 1
-            elif len(self.gene_info.all_isoforms_exons[refrenence_isoform_id]) > 1:
-                if self.read_assignment_counts[read_assignment.read_id] > 0:
-                    pass
-                    # logger.debug("Spliced read %s was previously used for construction, assigned id %s" %
-                    #             (read_assignment.read_id, refrenence_isoform_id))
-                spliced_isoform_reads[refrenence_isoform_id].append(read_assignment)
-
-                if self.args.requires_polya_for_construction and self.gene_info.isoform_strands[refrenence_isoform_id] == '-':
-                    if any(x.event_type == MatchEventSubtype.correct_polya_site_left for x in events):
-                        isoform_left_support[refrenence_isoform_id] += 1
-                elif abs(self.gene_info.all_isoforms_exons[refrenence_isoform_id][0][0] - read_assignment.corrected_exons[0][0]) <= self.args.apa_delta:
-                    isoform_left_support[refrenence_isoform_id] += 1
-
-                if self.args.requires_polya_for_construction and self.gene_info.isoform_strands[refrenence_isoform_id] == '+':
-                    if any(x.event_type == MatchEventSubtype.correct_polya_site_right for x in events):
-                        isoform_right_support[refrenence_isoform_id] += 1
-                elif abs(self.gene_info.all_isoforms_exons[refrenence_isoform_id][-1][1] - read_assignment.corrected_exons[-1][1]) <= self.args.apa_delta:
-                    isoform_right_support[refrenence_isoform_id] += 1
-
-        self.construct_monoexon_isoforms(mono_exon_isoform_reads, mono_exon_isoform_coverage, polya_sites)
-        if not self.args.fl_only:
-            logger.debug("Constructing nonFL isoforms")
-            self.construct_nonfl_isoforms(spliced_isoform_reads, isoform_left_support, isoform_right_support)
-        if self.args.report_novel_unspliced:
-            self.construct_monoexon_novel(novel_mono_exon_reads)
-
-    def collect_terminal_exons_from_graph(self):
-        polya_exons = []
-        polyt_exons = []
-        for intron in self.intron_graph.outgoing_edges.keys():
-            for v in self.intron_graph.outgoing_edges[intron]:
-                if v[0] == TerminalVertex.polya:
-                    polya_exons.append((intron[1], v[1]))
-        for intron in self.intron_graph.incoming_edges.keys():
-            for v in self.intron_graph.incoming_edges[intron]:
-                if v[0] == TerminalVertex.polyt:
-                    polyt_exons.append((v[1], intron[0]))
-        # logger.debug("PolyA terminal exons: " + str(polya_exons))
-        # logger.debug("PolyT terminal exons: " + str(polyt_exons))
-        return polya_exons, polyt_exons
-
-    def is_internal_monoexonic_read(self, alignment, terminal_exons, forward=True):
-        read_coordinates = alignment.corrected_exons[0]
-        if forward:
-            for e in terminal_exons:
-                if abs(e[1] - alignment.corrected_exons[-1][1]) <= self.args.apa_delta and \
-                        read_coordinates[0] >= e[0] - self.args.delta:
-                    return True
-        else:
-            for e in terminal_exons:
-                if abs(e[0] - alignment.corrected_exons[0][0]) <= self.args.apa_delta and \
-                        read_coordinates[1] <= e[1] + self.args.delta:
-                    return True
-        return False
-
-    def construct_monoexon_novel(self, novel_mono_exon_reads):
-        logger.debug("Constructing novel monoexon")
-        polya_exons, polyt_exons = self.collect_terminal_exons_from_graph()
-        polya_reads = defaultdict(list)
-        polyt_reads = defaultdict(list)
-        for a in novel_mono_exon_reads:
-            if a.polya_info.external_polya_pos != -1:
-                if not self.is_internal_monoexonic_read(a, polya_exons, forward=True):
-                    polya_reads[a.polya_info.external_polya_pos].append(a)
-            if a.polya_info.external_polyt_pos != -1:
-                if not self.is_internal_monoexonic_read(a, polyt_exons, forward=False):
-                    polyt_reads[a.polya_info.external_polyt_pos].append(a)
-
-        novel_monoexon = set()
-        clustered_polya_reads = self.cluster_monoexons(polya_reads)
-        novel_monoexon.update(self.generate_monoexon_from_clustered(clustered_polya_reads, True))
-        clustered_polyt_reads = self.cluster_monoexons(polyt_reads)
-        novel_monoexon.update(self.generate_monoexon_from_clustered(clustered_polyt_reads, False))
-
-    def generate_monoexon_from_clustered(self, clustered_reads, forward=True):
-        cutoff = self.args.min_novel_count
-        result = set()
-        for three_prime_pos in clustered_reads.keys():
-            count = len(clustered_reads[three_prime_pos])
-            if count < cutoff:
-                continue
-            # TODO improve
-            if forward:
-                five_prime_pos = min([a.corrected_exons[0][0] for a in clustered_reads[three_prime_pos]])
-            else:
-                five_prime_pos = max([a.corrected_exons[-1][1] for a in clustered_reads[three_prime_pos]])
-
-            strand = '+' if forward else '-'
-            coordinates = (five_prime_pos, three_prime_pos) if forward else (three_prime_pos, five_prime_pos)
-            new_transcript_id = TranscriptNaming.transcript_prefix + str(self.get_transcript_id())
-            transcript_gene = (TranscriptNaming.novel_gene_prefix + self.gene_info.chr_id +
-                               "_" + str(self.get_transcript_id()))
-            transcript_type = TranscriptModelType.novel_not_in_catalog
-            id_suffix = TranscriptNaming.nnic_transcript_suffix
-
-            is_valid = True
-            half_len = interval_len(coordinates) / 2
-            for existing_model in self.transcript_model_storage:
-                if any(intersection_len(exon, coordinates) > half_len for exon in existing_model.exon_blocks):
-                    is_valid = False
-                    break
-            if not is_valid:
-                continue
-
-            new_model = TranscriptModel(self.gene_info.chr_id, strand,
-                                        new_transcript_id + ".%s" % self.gene_info.chr_id + id_suffix,
-                                        transcript_gene, [coordinates], transcript_type)
-            # logger.debug("uuu Adding novel MONOEXON isoform %s : %s, %d\t%d" % (new_transcript_id, str(coordinates), count, cutoff))
-            result.add(coordinates)
-
-            self.transcript_model_storage.append(new_model)
-            for read_assignment in clustered_reads[three_prime_pos]:
-                self.save_assigned_read(read_assignment, new_model.transcript_id)
-        return result
-
-    def cluster_monoexons(self, grouped_reads):
-        clustered_counts = defaultdict(list)
-        while grouped_reads:
-            best_pair = max(grouped_reads.items(), key=lambda x:len(x[1]))
-            top_position = best_pair[0]
-            for pos in range(top_position - self.args.apa_delta, top_position + self.args.apa_delta + 1):
-                if pos in grouped_reads:
-                    clustered_counts[top_position] += grouped_reads[pos]
-                    del grouped_reads[pos]
-
-        return clustered_counts
-
-    def construct_monoexon_isoforms(self, mono_exon_isoform_reads, mono_exon_isoform_coverage, polya_sites):
-        for isoform_id in mono_exon_isoform_reads.keys():
-            count = len(mono_exon_isoform_reads[isoform_id])
-            coverage = float(mono_exon_isoform_coverage[isoform_id].count(1)) / \
-                       float(len(mono_exon_isoform_coverage[isoform_id]))
-            polya_support = polya_sites[isoform_id]
-
-            # logger.debug(">> Monoexon transcript %s: %d\t%d\t%.4f\t%d" % (isoform_id, self.intron_graph.max_coverage, count, coverage, polya_support))
-            if (count < self.args.min_known_count or coverage < self.args.min_mono_exon_coverage or
-                    (self.args.require_monoexonic_polya and polya_support == 0)):
-                pass # logger.debug(">> Will NOT be added, abs cutoff=%d" % (self.params.min_known_count))
-            elif isoform_id not in GraphBasedModelConstructor.detected_known_isoforms:
-                new_model = self.transcript_from_reference(isoform_id)
-                self.transcript_model_storage.append(new_model)
-                GraphBasedModelConstructor.detected_known_isoforms.add(isoform_id)
-                for read_assignment in mono_exon_isoform_reads[isoform_id]:
-                    self.save_assigned_read(read_assignment, new_model.transcript_id)
-                # logger.debug(">> Adding known monoexon isoform %s, %s, count = %d: %s" %
-                #             (self.transcript_model_storage[-1].transcript_id, isoform_id,
-                #              count, str(self.gene_info.all_isoforms_exons[isoform_id])))
-
-    def construct_nonfl_isoforms(self, spliced_isoform_reads, spliced_isoform_left_support, spliced_isoform_right_support):
-        logger.debug("Constructing nonFL isoforms")
-        for isoform_id in spliced_isoform_reads.keys():
-            if isoform_id in GraphBasedModelConstructor.detected_known_isoforms:
-                continue
-            count = len(spliced_isoform_reads[isoform_id])
-            if isoform_id not in self.known_isoforms_in_graph_ids:
-                #logger.debug("<< Isoform %s has %d assignments but is not in the graph" % (isoform_id, count))
-                continue
-
-            intron_path = self.known_isoforms_in_graph_ids[isoform_id]
-            #logger.debug("Known non-FL spliced isoform %s" % isoform_id)
-            if count < self.args.min_known_count or \
-                    spliced_isoform_left_support[isoform_id] < 1 or \
-                    spliced_isoform_right_support[isoform_id] < 1:
-                pass
-                # logger.debug("Will not be added")
-            else:
-                logger.debug("<< Adding known non-FL spliced isoform %s" % isoform_id)
-                new_model = self.transcript_from_reference(isoform_id)
-                self.transcript_model_storage.append(new_model)
-                GraphBasedModelConstructor.detected_known_isoforms.add(isoform_id)
-                # Remember which reads built this known so that, after filtering,
-                # resolve_read_ownership_conflicts can hand back any read still
-                # shared with a surviving novel FL model.
-                self.nonfl_known_reads[new_model.transcript_id] = list(spliced_isoform_reads[isoform_id])
-                for read_assignment in spliced_isoform_reads[isoform_id]:
-                    self.save_assigned_read(read_assignment, new_model.transcript_id)
-
-    # create transcript model object from reference isoforms
-    def transcript_from_reference(self, isoform_id):
-        return TranscriptModel.from_reference_transcript(self.gene_info, isoform_id)
-
-    @staticmethod
-    def _reference_unique_known_isoform(assignment: ReadAssignment) -> Optional[str]:
-        # The known reference isoform this read was uniquely assigned to by the
-        # reference step, or None. All unique reference assignments target an
-        # annotated (known) isoform, so the returned id also names the model
-        # built from it (from_reference_transcript keeps the reference id).
-        if assignment.assignment_type.is_unique() and assignment.isoform_matches:
-            return assignment.isoform_matches[0].assigned_transcript
-        return None
-
-    # assign reads back to constructed isoforms
-    def assign_reads_to_models(self, read_assignments):
-        if not self.transcript_model_storage:
-            for assignment in read_assignments:
-                read_id = assignment.read_id
-                self.read_assignment_counts[read_id] = 0
-            logger.debug("No transcripts were assigned")
-            return
-
-        logger.debug("Creating artificial GeneInfo from %d transcript models" % len(self.transcript_model_storage))
-        transcript_model_gene_info = GeneInfo.from_models(self.transcript_model_storage, self.args.delta)
-        assigner = LongReadAssigner(transcript_model_gene_info, self.args, self.string_pools, quick_mode=True)
-        profile_constructor = CombinedProfileConstructor(transcript_model_gene_info, self.args)
-        # ids of the isoforms actually built as models here; a known reference
-        # isoform keeps its id (from_reference_transcript), so this doubles as
-        # the set of surviving known isoforms.
-        model_ids = set(transcript_model_gene_info.all_isoforms_exons.keys())
-        kept_on_known = 0
-        dropped_no_model = 0
-
-        for assignment in read_assignments:
-            read_id = assignment.read_id
-            if self.read_assignment_counts[read_id] > 0:
-                # logger.debug("# Read %s was assigned to based on path construction" % read_id)
-                continue
-
-            # Consistency gate: a read the reference step assigned uniquely to a
-            # known isoform stays on that isoform and is never re-profiled onto a
-            # different model. If that isoform was not built as a model (filtered
-            # out / unsupported), the read is left unassigned rather than dragged
-            # elsewhere. Inert without --genedb (no known reference isoforms).
-            known_iso = self._reference_unique_known_isoform(assignment)
-            if known_iso is not None:
-                if known_iso in model_ids:
-                    self.transcript_read_ids[known_iso].append(assignment)
-                    self.internal_counter[known_iso] += 1
-                    self.read_assignment_counts[read_id] += 1
-                    kept_on_known += 1
-                else:
-                    self.read_assignment_counts[read_id] = 0
-                    dropped_no_model += 1
-                continue
-
-            read_exons = assignment.corrected_exons
-            # logger.debug("# Checking read %s: %s" % (assignment.read_id, str(read_exons)))
-            model_combined_profile = profile_constructor.construct_profiles(read_exons, assignment.polya_info, [])
-            model_assignment = assigner.assign_to_isoform(assignment.read_id, model_combined_profile)
-            model_assignment.read_group = assignment.read_group  # Full list, not just [0]
-            # check that no serious contradiction occurs
-            if model_assignment.assignment_type.is_consistent():
-                matched_isoforms = [m.assigned_transcript for m in model_assignment.isoform_matches]
-
-                if len(matched_isoforms) == 1:
-                    self.internal_counter[matched_isoforms[0]] += 1
-                for m in model_assignment.isoform_matches:
-                    self.read_assignment_counts[read_id] += 1
-                    self.transcript_read_ids[m.assigned_transcript].append(assignment)
-            else:
-                self.read_assignment_counts[read_id] = 0
-
-        logger.debug("Gated re-assignment: kept %d reads on their unique known isoform, "
-                     "dropped %d (known isoform not built)" % (kept_on_known, dropped_no_model))
-
-    def correct_novel_transcript_ends(self, transcript_model: TranscriptModel,
-                                      assigned_reads: List[ReadAssignment]) -> None:
-        # Part 1: detector-based per-transcript end refinement. Build read-terminus
-        # histograms from this model's own reads and snap each unsupported terminal
-        # exon end to the dominant confident polyA/TSS peak (XGBoost detect_peaks);
-        # both ends are refined in one pass over the same reads so the chosen
-        # 5'/3' stay concordant. polyA always; TSS only when use_tss_model
-        # (--fl_data). No confident peak / no model -> positional fallback, then
-        # leave as is. Never creates or drops a model.
-        logger.debug("Verifying ends for transcript %s" % transcript_model.transcript_id)
-        transcript_start = transcript_model.exon_blocks[0][0]
-        transcript_end = transcript_model.exon_blocks[-1][1]
-        first_exon_right = transcript_model.exon_blocks[0][1]
-        last_exon_left = transcript_model.exon_blocks[-1][0]
-        strand = transcript_model.strand
-
-        start_reads, end_reads = self._terminal_histograms(transcript_model, assigned_reads)
-        # _terminal_histograms maps position -> reads; detect_peaks / _peak_boundary and
-        # _closest_inward want position -> count, so collapse the read lists to counts
-        # (same as derive_alternative_end_models).
-        start_hist = {p: len(rs) for p, rs in start_reads.items()}
-        end_hist = {p: len(rs) for p, rs in end_reads.items()}
-        start_supported = any(abs(p - transcript_start) <= self.args.apa_delta for p in start_hist)
-        end_supported = any(abs(p - transcript_end) <= self.args.apa_delta for p in end_hist)
-
-        if not start_supported:
-            model = self._terminal_model(strand, left=True)
-            if model is not None:
-                new_start = self._peak_boundary(start_hist, model, lambda pos: pos < first_exon_right)
-            else:
-                new_start = self._closest_inward(sorted(start_hist.keys()), transcript_start, greater=True)
-            if new_start is not None and new_start != transcript_start and new_start < first_exon_right:
-                logger.debug("Changed start for transcript %s: from %d to %d" %
-                             (transcript_model.transcript_id, transcript_start, new_start))
-                transcript_model.exon_blocks[0] = (new_start, first_exon_right)
-
-        if not end_supported:
-            model = self._terminal_model(strand, left=False)
-            if model is not None:
-                new_end = self._peak_boundary(end_hist, model, lambda pos: pos > last_exon_left)
-            else:
-                new_end = self._closest_inward(sorted(end_hist.keys(), reverse=True), transcript_end, greater=False)
-            if new_end is not None and new_end != transcript_end and new_end > last_exon_left:
-                logger.debug("Changed end for transcript %s: from %d to %d" %
-                             (transcript_model.transcript_id, transcript_end, new_end))
-                transcript_model.exon_blocks[-1] = (last_exon_left, new_end)
-
-    def _mark_terminal_confirmation(self, model: TranscriptModel) -> None:
-        # Record whether the model's polyA (3') / TSS (5') terminus is supported by a
-        # read-terminus peak among its own reads, using the same histograms as
-        # correct_novel_transcript_ends (the 3' side counts polyA-confirmed reads only).
-        # Called after end refinement so the check uses the model's final ends.
-        reads = self.transcript_read_ids[model.transcript_id]
-        if not reads:
-            model.polya_confirmed = model.tss_confirmed = False
-            return
-        start_reads, end_reads = self._terminal_histograms(model, reads)
-        model_start = model.exon_blocks[0][0]
-        model_end = model.exon_blocks[-1][1]
-        start_supported = any(abs(p - model_start) <= self.args.apa_delta for p in start_reads)
-        end_supported = any(abs(p - model_end) <= self.args.apa_delta for p in end_reads)
-        if model.strand == '+':      # genomic-left = 5' TSS, genomic-right = 3' polyA
-            model.tss_confirmed, model.polya_confirmed = start_supported, end_supported
-        elif model.strand == '-':    # reversed
-            model.polya_confirmed, model.tss_confirmed = start_supported, end_supported
-        else:                        # unstranded: no polyA orientation
-            model.polya_confirmed = model.tss_confirmed = False
-
-    def _terminal_model(self, strand: str, left: bool) -> Optional["XGBClassifier"]:
-        # Which trained model applies to a genomic-side boundary. For '+' the
-        # right end is the polyA site and the left is the TSS; for '-' reversed.
-        # Returns None (-> positional fallback) for unstranded, or a side whose
-        # model is disabled: polyA needs the flag, TSS additionally needs fl_data.
-        if strand == '+':
-            is_polya_side = not left
-        elif strand == '-':
-            is_polya_side = left
-        else:
-            return None
-        if is_polya_side:
-            # polyA end refinement is always on (polishing needs no annotation).
-            return get_polya_model()
-        return get_tss_model() if self.use_tss_model else None
-
-    @staticmethod
-    def _peak_boundary(histogram: Dict[int, int], model: "XGBClassifier",
-                       valid: Callable[[int], bool]) -> Optional[int]:
-        # Best-supported detected peak satisfying the terminal-exon clamp.
-        if not histogram:
-            return None
-        peaks = [p for p in detect_peaks(histogram, model) if valid(p.position)]
-        if not peaks:
-            return None
-        return max(peaks, key=lambda p: p.count).position
-
-    @staticmethod
-    def _closest_inward(sorted_positions: List[int], boundary: int, greater: bool) -> Optional[int]:
-        # Positional fallback: nearest read terminus on the inward side of the
-        # current boundary (matches the original end-correction behaviour).
-        for pos in sorted_positions:
-            if greater and pos > boundary:
-                return pos
-            if not greater and pos < boundary:
-                return pos
-        return None
-
-    @staticmethod
-    def _confirmed_polya_pos(assignment: ReadAssignment, strand: str) -> Optional[int]:
-        # Detected polyA cleavage site of a read, matching the trained-model
-        # domain (PolyACounter): only polyA-confirmed reads, at external_polya_pos
-        # ('+') / external_polyt_pos ('-'). None otherwise.
-        info = getattr(assignment, 'polya_info', None)
-        if not getattr(assignment, 'polyA_found', False) or info is None:
-            return None
-        if strand == '+' and info.external_polya_pos != -1:
-            return info.external_polya_pos
-        if strand == '-' and info.external_polyt_pos != -1:
-            return info.external_polyt_pos
-        return None
-
-    def _terminal_histograms(self, source_model: TranscriptModel,
-                             assigned_reads: List[ReadAssignment]) \
-            -> Tuple[Dict[int, List[ReadAssignment]], Dict[int, List[ReadAssignment]]]:
-        # Build genomic-left (start) and genomic-right (end) read-terminus maps
-        # matching what each trained model was fit on:
-        #   - 3' polyA side: polyA-CONFIRMED reads at the detected cleavage site
-        #     (so a low overall polyA rate naturally raises the support bar);
-        #   - 5' TSS side: all stranded reads' alignment ends.
-        # Unstranded -> alignment ends on both sides (no polyA orientation).
-        # Each map is position -> the reads terminating there, so a peak's reads
-        # can later be moved onto the alternative-end model derived from it.
-        strand = source_model.strand
-        first_exon_right = source_model.exon_blocks[0][1]
-        last_exon_left = source_model.exon_blocks[-1][0]
-        start_reads = defaultdict(list)
-        end_reads = defaultdict(list)
-        for a in assigned_reads:
-            ex = a.corrected_exons
-            if strand == '+':
-                if ex[0][0] < first_exon_right:           # 5' TSS (all reads)
-                    start_reads[ex[0][0]].append(a)
-                pos = self._confirmed_polya_pos(a, '+')   # 3' polyA (confirmed)
-                if pos is not None and pos > last_exon_left:
-                    end_reads[pos].append(a)
-            elif strand == '-':
-                pos = self._confirmed_polya_pos(a, '-')   # 3' polyA (confirmed)
-                if pos is not None and pos < first_exon_right:
-                    start_reads[pos].append(a)
-                if ex[-1][1] > last_exon_left:            # 5' TSS (all reads)
-                    end_reads[ex[-1][1]].append(a)
-            else:
-                if ex[0][0] < first_exon_right:
-                    start_reads[ex[0][0]].append(a)
-                if ex[-1][1] > last_exon_left:
-                    end_reads[ex[-1][1]].append(a)
-        return start_reads, end_reads
-
-    @staticmethod
-    def _intron_chain_key(model: TranscriptModel) -> Tuple[Tuple[int, int], ...]:
-        # Terminal-end-independent identity of a transcript: the tuple of its
-        # introns derived from exon blocks, using the project-wide
-        # junctions_from_blocks convention (actual intron coordinates, +1/-1).
-        # Monoexon -> empty tuple.
-        return tuple(junctions_from_blocks(model.exon_blocks))
-
-    def _add_known_alternative_end_models(self) -> None:
-        # Part 2: for each known (reference) model, peak-call its own assigned
-        # reads and emit a NIC for every confident alternative polyA/TSS end,
-        # keeping the known (union). Deduplicated against everything already in
-        # storage (notably the graph-level NICs) by intron chain + both ends
-        # within apa_delta, so the same alternative end is never emitted twice.
-        existing_pairs = defaultdict(list)
-        # Seed with every reference isoform too: an alternative-end peak that
-        # lands on another annotated isoform of the same chain is that known, not
-        # a novel end, so it must be suppressed even if that reference was not
-        # emitted as a model in this locus.
-        for ref_exons in self.gene_info.all_isoforms_exons.values():
-            ck = tuple(junctions_from_blocks(ref_exons))
-            existing_pairs[ck].append((ref_exons[0][0], ref_exons[-1][1]))
-        for m in self.transcript_model_storage:
-            existing_pairs[self._intron_chain_key(m)].append(
-                (m.exon_blocks[0][0], m.exon_blocks[-1][1]))
-
-        new_models = []
-        for model in self.transcript_model_storage:
-            if model.transcript_type == TranscriptModelType.known:
-                pass
-            elif not self.novel_apa:
-                # Part 2 default: known transcripts only. Part 3 (--novel_apa)
-                # also spins off alternative-end siblings for novel chains.
-                continue
-            source_id = model.transcript_id
-            for nic, nic_reads in self.derive_alternative_end_models(
-                    model, self.transcript_read_ids[source_id]):
-                ck = self._intron_chain_key(nic)
-                ns, ne = nic.exon_blocks[0][0], nic.exon_blocks[-1][1]
-                if any(abs(ns - s) <= self.args.apa_delta and abs(ne - e) <= self.args.apa_delta
-                       for (s, e) in existing_pairs[ck]):
-                    continue
-                existing_pairs[ck].append((ns, ne))
-                new_models.append(nic)
-                # Hand the alternative-end peak's reads to the NIC so it is not a
-                # zero-read phantom; keep at least one read on the source.
-                for read in nic_reads:
-                    if len(self.transcript_read_ids[source_id]) <= 1:
-                        break
-                    self._move_read_to_model(read, source_id, nic.transcript_id)
-        if new_models:
-            logger.debug("Added %d known alternative-end NICs" % len(new_models))
-            self.transcript_model_storage.extend(new_models)
-
-    def _drop_duplicate_alt_end_models(self) -> None:
-        # Final dedup over the whole storage (catches graph-level NICs too): drop
-        # any non-known model that is structurally identical (intron chain + both
-        # ends within apa_delta) to a reference isoform -> it is that annotated
-        # transcript, not a novel end; and collapse non-known models that
-        # duplicate an already-kept one.
-        ref_pairs = defaultdict(list)
-        for ref_exons in self.gene_info.all_isoforms_exons.values():
-            ck = tuple(junctions_from_blocks(ref_exons))
-            ref_pairs[ck].append((ref_exons[0][0], ref_exons[-1][1]))
-
-        d = self.args.apa_delta
-        kept = []
-        kept_pairs = defaultdict(list)
-        dropped = 0
-        for m in self.transcript_model_storage:
-            if m.transcript_type == TranscriptModelType.known:
-                kept.append(m)
-                continue
-            ck = self._intron_chain_key(m)
-            s, e = m.exon_blocks[0][0], m.exon_blocks[-1][1]
-            if any(abs(s - rs) <= d and abs(e - re) <= d for rs, re in ref_pairs[ck]) or \
-                    any(abs(s - ks) <= d and abs(e - ke) <= d for ks, ke in kept_pairs[ck]):
-                dropped += 1
-                # free the model's reads for reassignment (graph-level NICs are
-                # already counted; post-pass NICs are not yet in the counters)
-                if m.transcript_id in self.internal_counter:
-                    self.delete_from_storage(m.transcript_id)
-                continue
-            kept.append(m)
-            kept_pairs[ck].append((s, e))
-        if dropped:
-            logger.debug("Dropped %d duplicate / reference-matching alt-end models" % dropped)
-        self.transcript_model_storage = kept
-
-    def derive_alternative_end_models(self, source_model: TranscriptModel,
-                                      assigned_reads: List[ReadAssignment]) \
-            -> List[Tuple[TranscriptModel, List[ReadAssignment]]]:
-        # Confident alternative polyA/TSS ends for a transcript, from its own
-        # reads (per-transcript, so 5'/3' stay concordant). Returns (new model,
-        # supporting reads) per alternative end, each model changing a single
-        # terminal. The supporting reads are the source reads whose terminus sits
-        # at that peak (within apa_delta); the caller moves them onto the NIC so it
-        # is not a zero-read phantom. Each read backs at most one NIC. polyA always;
-        # TSS only when use_tss_model (--fl_data). A known source yields
-        # novel-in-catalog siblings; a novel source keeps its own type.
-        if not assigned_reads:
-            return []
-        exon_blocks = source_model.exon_blocks
-        strand = source_model.strand
-        annotated_start = exon_blocks[0][0]
-        annotated_end = exon_blocks[-1][1]
-        first_exon_right = exon_blocks[0][1]
-        last_exon_left = exon_blocks[-1][0]
-        sibling_type = (TranscriptModelType.novel_in_catalog
-                        if source_model.transcript_type == TranscriptModelType.known
-                        else source_model.transcript_type)
-
-        start_reads, end_reads = self._terminal_histograms(source_model, assigned_reads)
-        start_hist = {p: len(rs) for p, rs in start_reads.items()}
-        end_hist = {p: len(rs) for p, rs in end_reads.items()}
-
-        claimed = set()  # id(read) already routed to a NIC (one NIC per read, both sides)
-
-        def assign_to_peaks(reads_map, alt_peaks, annotated_pos):
-            # Route each read to the *nearest* terminus among the annotated end and
-            # the alternative-end peaks (within apa_delta). Reads nearest the
-            # annotated end stay on the source; reads nearest an alt peak back that
-            # peak's NIC. Nearest (not first-within-window) avoids a peak greedily
-            # claiming a neighbour's reads and starving it; the shared claimed set
-            # keeps a both-ends-alternative read on a single NIC (it is moved once).
-            per_peak = {p: [] for p in alt_peaks}
-            candidates = list(alt_peaks) + [annotated_pos]
-            for p, rs in reads_map.items():
-                nearest = min(candidates, key=lambda c: (abs(c - p), c))
-                if nearest != annotated_pos and abs(nearest - p) <= self.args.apa_delta:
-                    for r in rs:
-                        if id(r) not in claimed:
-                            claimed.add(id(r))
-                            per_peak[nearest].append(r)
-            return per_peak
-
-        new_models = []
-        start_model = self._terminal_model(strand, left=True)
-        if start_model is not None:
-            start_peaks = self._alternative_end_positions(start_hist, start_model, annotated_start,
-                                                          lambda p: p < first_exon_right)
-            per_peak = assign_to_peaks(start_reads, start_peaks, annotated_start)
-            for pos in start_peaks:
-                nic = self._nic_model_with_boundary(source_model, sibling_type, start=pos)
-                new_models.append((nic, per_peak[pos]))
-        end_model = self._terminal_model(strand, left=False)
-        if end_model is not None:
-            end_peaks = self._alternative_end_positions(end_hist, end_model, annotated_end,
-                                                        lambda p: p > last_exon_left)
-            per_peak = assign_to_peaks(end_reads, end_peaks, annotated_end)
-            for pos in end_peaks:
-                nic = self._nic_model_with_boundary(source_model, sibling_type, end=pos)
-                new_models.append((nic, per_peak[pos]))
-        return new_models
-
-    def _alternative_end_positions(self, histogram: Dict[int, int], model: "XGBClassifier",
-                                   annotated_pos: int, clamp: Callable[[int], bool]) -> List[int]:
-        # Detected peaks representing a genuine alternative end: inside the
-        # terminal exon, > apa_delta from the annotated end, and clearing BOTH an
-        # absolute (min_novel_count) and a RELATIVE support cutoff. The relative
-        # cutoff (terminal_position_rel x the transcript's dominant terminal peak,
-        # reusing the intron-graph terminal threshold) rejects minor secondary
-        # peaks that are not real co-expressed isoforms and self-normalizes per
-        # transcript, so it does not depend on absolute depth or fit one isoform.
-        if not histogram:
-            return []
-        peaks = [p for p in detect_peaks(histogram, model) if clamp(p.position)]
-        if not peaks:
-            return []
-        dominant = max(p.count for p in peaks)
-        cutoff = max(self.args.min_novel_count, self.args.terminal_position_rel * dominant)
-        return [p.position for p in peaks
-                if abs(p.position - annotated_pos) > self.args.apa_delta and p.count >= cutoff]
-
-    def _nic_model_with_boundary(self, source_model: TranscriptModel, transcript_type: TranscriptModelType,
-                                 start: Optional[int] = None, end: Optional[int] = None) -> TranscriptModel:
-        # Build an alternative-end model from a source one by replacing a single
-        # terminal coordinate; the intron chain (internal exons) is unchanged.
-        # transcript_type is novel_in_catalog for a known source, or the source's
-        # own type for a novel source.
-        exon_blocks = [tuple(e) for e in source_model.exon_blocks]
-        if start is not None:
-            exon_blocks[0] = (start, exon_blocks[0][1])
-        if end is not None:
-            exon_blocks[-1] = (exon_blocks[-1][0], end)
-        suffix = (TranscriptNaming.nic_transcript_suffix
-                  if transcript_type == TranscriptModelType.novel_in_catalog
-                  else TranscriptNaming.nnic_transcript_suffix)
-        new_transcript_id = TranscriptNaming.transcript_prefix + str(self.get_transcript_id())
-        new_model = TranscriptModel(
-            self.gene_info.chr_id, source_model.strand,
-            new_transcript_id + ".%s" % self.gene_info.chr_id + suffix,
-            source_model.gene_id, exon_blocks, transcript_type)
-        new_model.intron_path = source_model.intron_path
-        logger.debug("Adding alternative-end model %s from %s (start=%s, end=%s)" %
-                     (new_model.transcript_id, source_model.transcript_id, str(start), str(end)))
-        return new_model
-
-
-class IntronPathStorage:
-    def __init__(self, params, path_processor):
-        self.params = params
-        self.path_processor = path_processor
-        self.intron_graph = path_processor.intron_graph
-        self.paths = defaultdict(int)
-        self.fl_paths = set()
-        self.paths_to_reads = defaultdict(list)
-
-    def fill(self, read_assignments):
-        for a in read_assignments:
-            if a.multimapper:
-                continue
-            intron_path = self.path_processor.thread_introns(a.corrected_introns)
-            if not intron_path:
-                continue
-            read_end = a.corrected_exons[-1][1]
-            is_end_trusted = a.strand == '+' and \
-                             (a.polya_info.external_polya_pos != -1 or
-                              a.polya_info.internal_polya_pos != -1)
-            terminal_vertex = self.path_processor.thread_ends(intron_path[-1], read_end, is_end_trusted)
-            if terminal_vertex:
-                intron_path.append(terminal_vertex)
-
-            read_start = a.corrected_exons[0][0]
-            is_start_trusted = a.strand == '-' and \
-                               (a.polya_info.external_polyt_pos != -1 or
-                                a.polya_info.internal_polyt_pos != -1)
-            starting_vertex = self.path_processor.thread_starts(intron_path[0], read_start, is_start_trusted)
-            if starting_vertex:
-                intron_path = [starting_vertex] + intron_path
-
-            path_tuple = tuple(intron_path)
-            self.paths[path_tuple] += 1
-            if terminal_vertex and starting_vertex:
-                if not self.params.requires_polya_for_construction or\
-                        (terminal_vertex[0] == TerminalVertex.polya or starting_vertex[0] == TerminalVertex.polyt):
-                    self.fl_paths.add(path_tuple)
-            self.paths_to_reads[path_tuple].append(a)
-
-
-class IntronPathProcessor:
-    def __init__(self, params, intron_graph):
-        self.params = params
-        self.intron_graph = intron_graph
-        self.all_vertices = set()
-        self.all_vertices.update(self.intron_graph.intron_collector.clustered_introns.keys())
-        for edge_set in self.intron_graph.outgoing_edges.values():
-            self.all_vertices.update(edge_set)
-        for edge_set in self.intron_graph.incoming_edges.values():
-            self.all_vertices.update(edge_set)
-        self.visited = set()
-
-    def visit_vertex(self, v):
-        if v in self.all_vertices:
-            self.visited.add(v)
-
-    def visit_path(self, p):
-        for v in p:
-            self.visit_vertex(v)
-
-    def thread_introns(self, introns):
-        path = []
-        for intron in introns:
-            if intron in self.intron_graph.intron_collector.discarded_introns:
-                return None
-            path.append(self.intron_graph.intron_collector.substitute(intron))
-        return path
-
-    def thread_ends(self, intron, end, trusted=False):
-        possible_polyas = self.intron_graph.get_outgoing(intron, TerminalVertex.polya)
-        if trusted:
-            # find closes polyA
-            for v in possible_polyas:
-                if abs(v[1] - end) <= self.params.apa_delta:
-                    return v
-
-        outgoing_introns = self.intron_graph.get_outgoing(intron)
-        if len(outgoing_introns) > 0:
-            # intron has outgoing edges
-            rightmost_exon_end = max([intron[0] for intron in outgoing_introns]) - 1
-            if not trusted and end <= rightmost_exon_end + self.params.delta:
-                # read end lies within next exon and has no polyA
-                return None
-
-        # consider all terminal position available for intron (bare read ends,
-        # confirmed 5' TSS ends for '-' transcripts, and confirmed polyA ends)
-        all_possible_ends = sorted(list(self.intron_graph.get_outgoing(intron, TerminalVertex.read_end)) +
-                                   list(self.intron_graph.get_outgoing(intron, TerminalVertex.tss_right)) +
-                                   list(possible_polyas), key=lambda x:x[1])
-        if len(all_possible_ends) == 0:
-            return None
-
-        rightmost_end = all_possible_ends[-1]
-        if trusted and end >= rightmost_end[1] and rightmost_end[0] == TerminalVertex.read_end:
-            # if we have trusted read, in cannot stop earlier that rightmost end (otherwise it should match polyA)
-            return rightmost_end
-        elif not trusted and end <= rightmost_end[1] + self.params.apa_delta and \
-                (len(all_possible_ends) <= 1 or end > all_possible_ends[-2][1]):
-            # non trusted should end before rightmost position + apa_delta but not earlier than second last
-            return rightmost_end
-        return None
-
-    def thread_starts(self, intron, start, trusted=False):
-        possible_polyas = self.intron_graph.get_incoming(intron, TerminalVertex.polyt)
-        if trusted:
-            # find closes polyT
-            for v in possible_polyas:
-                if abs(v[1] - start) <= self.params.apa_delta:
-                    return v
-
-        incoming_introns = self.intron_graph.get_incoming(intron)
-        if len(incoming_introns) > 0:
-            # intron has outgoing edges
-            leftmost_exon_start = min([intron[1] for intron in incoming_introns]) + 1
-            if not trusted and start >= leftmost_exon_start - self.params.delta:
-                # read start lies within previous exon and has no polyA
-                return None
-
-        all_possible_starts = sorted(list(self.intron_graph.get_incoming(intron, TerminalVertex.read_start)) +
-                                     list(self.intron_graph.get_incoming(intron, TerminalVertex.tss_left)) +
-                                     list(possible_polyas), key=lambda x: x[1])
-        if len(all_possible_starts) == 0:
-            return None
-
-        leftmost_start = all_possible_starts[0]
-        if trusted and start <= leftmost_start[1] and leftmost_start[0] == TerminalVertex.read_start:
-            return leftmost_start
-        elif not trusted and start >= leftmost_start[1] and \
-                (len(all_possible_starts) <= 1 or start < all_possible_starts[1][1]):
-            return leftmost_start
-        return None
