+++ isoquant_lib/model_construction/__init__.py
@@ -1,0 +1,9 @@
+############################################################################
+# Copyright (c) 2022-2026 University of Helsinki
+# # All Rights Reserved
+# See file LICENSE for details.
+############################################################################
+
+# Transcript model construction, split out of the former monolithic
+# graph_based_model_construction.GraphBasedModelConstructor. See
+# .claude/ plan notes for the stage decomposition.
